refactor: enum 구조 변경
