+++ src/main/java/lotto/domain/constants/message/InputError.java
@@ -1,6 +1,6 @@
-package lotto.domain.errors;
+package lotto.domain.constants.message;
 
-public enum InputError implements Errors{
+public enum InputError implements Message {
     DUPLICATE_LOTTO_NUMBER("[ERROR] 로또의 번호 조합에 중복이 숫자가 발견되었습니다.");
 
     private final String message;
