+++ src/main/java/lotto/domain/constants/message/Message.java
@@ -1,0 +1,7 @@
+package lotto.domain.constants.message;
+
+public interface Message {
+
+    String getMessage();
+
+}
