+++ src/main/java/lotto/domain/constants/value/LottoRule.java
@@ -1,4 +1,4 @@
-package lotto.domain.constants;
+package lotto.domain.constants.value;
 
 public enum LottoRule {
     COMBINATION_LENGTH(6),
