--- src/main/java/lotto/domain/errors/Errors.java
@@ -1,7 +1,0 @@
-package lotto.domain.errors;
-
-public interface Errors {
-
-    String getMessage();
-
-}
